refactor(core): add CoreAPI class
To enhance the code readability, modify the composeAPI to return an
CoreAPI instance instead of an object.
