+++ packages/core/src/index.ts
@@ -1,3 +1,4 @@
 /** @module core */
 
+export * from './CoreAPI';
 export * from './composeAPI';
